+++ .gitignore
@@ -1,2 +1,4 @@
 # .gitignore
-node_modules/
+node_modules/*
+
+!luxon
