+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="index.css">
     <title>Awesome Books</title>
 </head>
 <body>
@@ -61,6 +61,7 @@
     <footer>
             <h2>Copyright...</h2>
     </footer>
-    <script src="app.js"></script>
+    <script type="module" src="index.js"></script>
+    <script src="./modules/displayScreen.js"></script>
 </body>
 </html>
