+++ index.js
@@ -1,0 +1,8 @@
+/* eslint-disable linebreak-style */
+/* eslint-disable max-classes-per-file */
+import { Book } from './modules/classBook.js';
+import { Store } from './modules/classStore.js';
+import { Interactions } from './modules/classInteractions.js';
+import { displayBookEvent, addBookEvent, removeBookEvent } from './modules/interactionEvents.js';
+import { input } from './modules/submitOnEnter.js';
+import { displayDate } from './node_modules/luxon/src/luxon.js';
