+++ modules/luxon.js
@@ -1,0 +1,2 @@
+export const now = DateTime.now();
+export const displayDate = document.getElementById('date').innerHTML = now;
