+++ modules/submitOnEnter.js
@@ -1,0 +1,7 @@
+export const input = document.getElementById('myInput');
+input.addEventListener('keyup', (event) => {
+  if (event.KeyboardEvent.code === 13) {
+    event.preventDefault();
+    document.getElementById('myBtn').click();
+  }
+});
